Share the sidebar's tab glyph with the switcher, and fix the text fallback's scale
Three preference fixes and one rendering fix, all in the switcher's cards.

The cards drew a hardcoded terminal symbol and their own spinner, so of the
four preferences behind a tab's icon only `showAgentIcons` was honoured:
"None" and every numbered or custom icon still rendered a terminal glyph,
the spinner appeared whether or not the status indicator was enabled, and
`showSpinnerOverAgentIcons`, the agent's brand color and the done dot were
all lost.

Rather than re-derive that in a second place, the decision now lives once in
`TabGlyph` (new `Macterm/Views/TabIcon.swift`), which both the sidebar row
and the switcher card render. `TabStatusGlyph`, `SidebarRowIcon` (now
`TabRowIcon`, since project rows use it too) and `NumberGlyph` move there
with it. This also collapses the sidebar's own duplicated branches — its two
Label icon builders were spelling the same rules twice.

`TabGlyph` resolves to no view at all when "None" leaves nothing to draw, so
neither caller reserves space for an invisible image: the sidebar's Label
keeps its icon column collapsed and the switcher's title stays flush with
its preview's left edge. It also takes a tint, so the selected card's icon
brightens with its title instead of staying dim beside white text (an agent
logo keeps its brand color — that is identity, not state).

Separately: a tab that has never been focused has no captured frame, so its
card falls back to the viewport text — and that fallback was typeset at a
fixed 5pt and bottom-aligned. In a ~100pt card an idle shell's prompt then
spanned the full width, several times larger than the same prompt in a
captured frame beside it, and sat at the bottom of a card whose real pane
draws it at the top. The fallback now maps the pane's own column count onto
the card's width, so a line lands at the scale a real thumbnail would have
shown, and rows run top down like the screen they came from. Verified
against a live pane: both paths now put the prompt at the top at ~25% of the
card's width.
